+++ lib/shrine/plugins/parallelize.rb
@@ -1,11 +1,9 @@
-require "thread/pool"
-
-Thread::Pool.abort_on_exception = true
+require "thread"
 
 class Shrine
   module Plugins
-    # The parallelize plugin parallelizes your uploads and deletes using the
-    # [thread] gem.
+    # The parallelize plugin parallelizes your uploads and deletes using
+    # threads.
     #
     #     plugin :parallelize
     #
@@ -17,8 +15,6 @@ class Shrine
     # By default a pool of 3 threads will be used, but you can change that:
     #
     #     plugin :parallelize, threads: 5
-    #
-    # [thread]: https://github.com/meh/ruby-thread
     module Parallelize
       def self.configure(uploader, threads: 3)
         uploader.opts[:parallelize_threads] = threads
@@ -36,19 +32,47 @@ class Shrine
         private
 
         def put(io, context)
-          context[:thread_pool].process { super }
+          context[:thread_pool].enqueue { super }
         end
 
         def remove(uploaded_file, context)
-          context[:thread_pool].process { super }
+          context[:thread_pool].enqueue { super }
         end
 
         # We initialize a thread pool with configured number of threads.
-        def with_pool
-          pool = Thread.pool(opts[:parallelize_threads])
+        def with_pool(&block)
+          pool = ThreadPool.new(opts[:parallelize_threads])
           result = yield pool
-          pool.shutdown
+          pool.perform
           result
+        end
+
+        class ThreadPool
+          def initialize(thread_count)
+            @thread_count = thread_count
+            @tasks = Queue.new
+          end
+
+          def enqueue(&task)
+            @tasks.enq(task)
+          end
+
+          def perform
+            threads = @thread_count.times.map { spawn_thread }
+            threads.each(&:join)
+          end
+
+          private
+
+          def spawn_thread
+            Thread.new do
+              Thread.current.abort_on_exception = true
+              loop do
+                task = @tasks.deq(true) rescue break
+                task.call
+              end
+            end
+          end
         end
       end
     end